+++ opentelemetry-sdk/CHANGELOG.md
@@ -9,6 +9,7 @@
   [#1066](https://github.com/open-telemetry/opentelemetry-rust/pull/1193).
 - Bump MSRV to 1.64 [#1203](https://github.com/open-telemetry/opentelemetry-rust/pull/1203)
 - Add unit/doc tests for MeterProvider #1220
+- Add in memory log exporter 
 
 ## v0.20.0